refactor: upgrade to v2
