+++ .github/workflows/ci.yml
@@ -24,7 +24,7 @@ jobs:
 
             - run: pnpm install
 
-            - run: pnpm -r run build
+            - run: pnpm run build
 
     eslint:
         runs-on: ubuntu-latest
@@ -42,7 +42,7 @@ jobs:
 
             - run: pnpm install
 
-            - run: pnpm -r run eslint
+            - run: pnpm run eslint
 
     typings:
         runs-on: ubuntu-latest
@@ -60,7 +60,7 @@ jobs:
 
             - run: pnpm install
 
-            - run: pnpm -r run tsc
+            - run: pnpm run tsc
 
     test:
         runs-on: ubuntu-latest
