+++ __tests__/interceptors.spec.ts
@@ -1,7 +1,6 @@
 import { defineMiddleware, fatcher } from '../src';
 import fetchMock from 'jest-fetch-mock';
-
-import { json } from '@fatcherjs/middleware-json';
+import { json } from './utils';
 
 const BASE_URL = 'https://fatcher.virtual';
