+++ __tests__/querystring.spec.ts
@@ -1,6 +1,6 @@
 import { fatcher, isFatcherError } from '../src';
 import fetchMock from 'jest-fetch-mock';
-import { json } from '@fatcherjs/middleware-json';
+import { json } from './utils';
 const BASE_URL = 'https://fatcher.virtual';
 
 describe('querystring', () => {
