+++ jest.config.ts
@@ -3,9 +3,8 @@ import { Config } from '@jest/types';
 const config: Config.InitialOptions = {
     testEnvironment: 'node',
     testMatch: ['**/?(*.)+(spec|test).[tj]s?(x)'],
-    roots: ['packages'],
     testPathIgnorePatterns: ['/node_modules/'],
-    collectCoverageFrom: ['packages/*/src/**/*.ts', '!packages/docs/**/*.ts', '!packages/mock/src/serviceWorker/*.ts'],
+    collectCoverageFrom: ['src/**/*.ts'],
     transform: {
         '^.+\\.(t|j)s$': 'ts-jest',
     },
