+++ package.json
@@ -1,42 +1,66 @@
 {
-    "private": true,
     "name": "fatcher",
     "version": "1.8.0",
+    "description": "Send fetch request gracefully in browsers and nodeJS",
+    "homepage": "https://fatcherjs.github.io/fatcher/",
     "repository": {
         "type": "git",
         "url": "git+https://github.com/fatcherjs/fatcher.git"
     },
+    "license": "MIT",
+    "keywords": [
+        "fetch",
+        "http",
+        "request",
+        "middlewares",
+        "streams",
+        "axios",
+        "got",
+        "node"
+    ],
+    "main": "dist/index.js",
+    "module": "dist/index.esm.js",
+    "browser": "dist/fatcher.min.js",
+    "unpkg": "dist/fatcher.min.js",
+    "typings": "dist/index.d.ts",
+    "engines": {
+        "node": ">=18.0.0"
+    },
+    "files": [
+        "dist"
+    ],
     "scripts": {
+        "dev": "rimraf dist && rollup -c rollup.config.ts -w",
+        "build": "rimraf dist && rollup -c rollup.config.ts",
+        "eslint": "eslint .",
+        "tsc": "tsc --noEmit",
         "test": "jest",
         "test:cov": "jest --coverage",
-        "prettier": "prettier -wu . --loglevel warn",
-        "changelog": "ts-node scripts/changelog.ts",
-        "prerelease": "pnpm run changelog && pnpm run prettier && ts-node scripts/prerelease.ts && pnpm -r run prerelease --if-present",
-        "ci": "pnpm -r run build && pnpm -r run eslint && pnpm -r run tsc && pnpm run test"
+        "deploy": "pnpm run build && pnpm publish --no-git-check",
+        "ci": "npm run eslint && npm run tsc && npm run build && npm run test"
+    },
+    "dependencies": {
+        "@fatcherjs/utils-shared": "^2.0.0"
     },
     "devDependencies": {
-        "@fansy/eslint-config": "^1.0.0",
+        "@fansy/eslint-config": "^1.1.0",
         "@fansy/prettier-config": "^1.0.0",
-        "@fatcherjs/middleware-json": "workspace:^",
-        "@jest/types": "^28.1.0",
-        "@rollup/plugin-node-resolve": "^13.1.3",
-        "@rollup/plugin-replace": "^4.0.0",
-        "@rollup/plugin-strip": "^2.1.0",
-        "@types/fs-extra": "^9.0.13",
-        "@types/jest": "^28.1.1",
-        "@types/node": "^18.0.0",
-        "esbuild": "^0.15.5",
-        "fast-glob": "^3.2.11",
-        "fs-extra": "^10.1.0",
-        "jest": "^28.1.0",
+        "@jest/types": "^29.2.0",
+        "@rollup/plugin-node-resolve": "^15.0.0",
+        "@rollup/plugin-replace": "^5.0.0",
+        "@rollup/plugin-strip": "^3.0.0",
+        "@types/jest": "^29.1.2",
+        "@types/node": "^18.11.0",
+        "esbuild": "^0.15.11",
+        "jest": "^29.2.0",
         "jest-fetch-mock": "^3.0.3",
         "rimraf": "^3.0.2",
-        "rollup": "^2.70.1",
+        "rollup": "^2.79.1",
         "rollup-plugin-bundle-size": "^1.0.3",
-        "rollup-plugin-dts": "^4.2.0",
-        "rollup-plugin-esbuild": "^4.8.2",
-        "ts-jest": "^28.0.2",
-        "ts-node": "^10.7.0",
-        "typescript": "^4.6.3"
+        "rollup-plugin-dts": "^4.2.3",
+        "rollup-plugin-esbuild": "^4.10.1",
+        "ts-jest": "^29.0.3",
+        "ts-node": "^10.9.1",
+        "typescript": "^4.8.4"
     }
 }
