--- packages/fatcher/README.md
@@ -1,62 +1,0 @@
-<div align="center">
-  <img alt="Fatcher logo" src="https://github.com/fatcherjs/fatcher/raw/master/packages/docs/public/fatcher.png" height="120" width="120" />
-</div>
-
-<div align="center">
-<h1>Fatcher</h1>
-
-Send fetch request gracefully in browsers and nodeJS
-
-[![codecov](https://codecov.io/gh/fatcherjs/fatcher/branch/master/graph/badge.svg?token=9DRTR2GXH8)](https://codecov.io/gh/fatcherjs/fatcher)
-[![](https://data.jsdelivr.com/v1/package/npm/fatcher/badge?style=rounded)](https://www.jsdelivr.com/package/npm/fatcher)
-[![install size](https://packagephobia.com/badge?p=fatcher)](https://packagephobia.com/result?p=fatcher)
-<a href="https://unpkg.com/fatcher"><img alt="Size" src="https://img.badgesize.io/https://unpkg.com/fatcher"></a>
-<a href="https://npmjs.com/package/fatcher"><img src="https://img.shields.io/npm/v/fatcher.svg" alt="npm package"></a>
-<a href="https://nodejs.org/en/about/releases/"><img src="https://img.shields.io/node/v/fatcher.svg" alt="node compatibility"></a>
-<a href="https://github.com/fatcherjs/fatcher/actions/workflows/ci.yml"><img src="https://github.com/fatcherjs/fatcher/actions/workflows/ci.yml/badge.svg?branch=master" alt="build status"></a>
-
-</div>
-
-## How it work
-
-<div align="center">
-  <img alt="Fatcher Workflow" src="https://github.com/fatcherjs/fatcher/raw/master/packages/docs/public/fatcher_workflow.png" />
-</div>
-
-## Why fetch
-
-<div align="center">
-  <img alt="Fatcher Workflow" src="https://github.com/fatcherjs/fatcher/raw/master/packages/docs/public/fatcher_streams.png" />
-</div>
-
-## Documentation
-
-➡️ [View documentation Here](https://fatcherjs.github.io/fatcher)
-
-## Install
-
-### NPM
-
-```bash
->$ npm install fatcher
-
-#or using yarn
->$ yarn add fatcher
-
-#or using pnpm
->$ pnpm add fatcher
-```
-
-### CDN
-
-```html
-<script src="https://cdn.jsdelivr.net/npm/fatcher/dist/fatcher.min.js"></script>
-```
-
-## Usage
-
-For detailed usage information,[visit the docs](https://fatcherjs.github.io/fatcher/quick-start).
-
-## License
-
-[MIT](https://github.com/fatcherjs/fatcher/blob/master/LICENSE)
