--- packages/fatcher/package.json
@@ -1,43 +1,0 @@
-{
-    "name": "fatcher",
-    "version": "1.8.0",
-    "description": "Send fetch request gracefully in browsers and nodeJS",
-    "homepage": "https://fatcherjs.github.io/fatcher/",
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/fatcherjs/fatcher.git"
-    },
-    "license": "MIT",
-    "keywords": [
-        "fetch",
-        "http",
-        "request",
-        "middlewares",
-        "streams",
-        "axios",
-        "got",
-        "node"
-    ],
-    "main": "dist/index.js",
-    "module": "dist/index.esm.js",
-    "browser": "dist/fatcher.min.js",
-    "unpkg": "dist/fatcher.min.js",
-    "typings": "dist/index.d.ts",
-    "engines": {
-        "node": ">=18.0.0"
-    },
-    "files": [
-        "dist"
-    ],
-    "scripts": {
-        "dev": "rimraf dist && rollup -c rollup.config.ts -w",
-        "build": "rimraf dist && rollup -c rollup.config.ts",
-        "eslint": "eslint ./src",
-        "tsc": "tsc --noEmit",
-        "deploy": "pnpm run build && pnpm publish --no-git-check",
-        "prerelease": "ts-node scripts/prerelease.ts"
-    },
-    "dependencies": {
-        "@fatcherjs/utils-shared": "^2.0.0"
-    }
-}
