--- packages/fatcher/tsconfig.json
@@ -1,4 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "include": ["src/**/*", "__tests__/**/*"]
-}
