--- shared/tests/getRandomString.ts
@@ -1,11 +1,0 @@
-export function getRandomString(length: number) {
-    const STRING_TEMPLATE = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890';
-
-    let string = '';
-
-    for (let i = 0; i < length; i++) {
-        string += STRING_TEMPLATE.charAt(Math.floor(Math.random() * STRING_TEMPLATE.length));
-    }
-
-    return string;
-}
