--- shared/tests/index.ts
@@ -1,4 +1,0 @@
-export * from './getRandomString';
-export * from './getStringStream';
-export * from './getStringStreamByLength';
-export * from './sleep';
