--- shared/tests/sleep.ts
@@ -1,3 +1,0 @@
-export function sleep(timeout = 500) {
-    return new Promise(resolve => setTimeout(resolve, timeout));
-}
