+++ src/errors/FatcherError.ts
