+++ src/errors/isFatcherError.ts
