+++ src/mergeHeaders.ts
