+++ src/options/mergeOptions.ts
