+++ tsconfig.json
@@ -9,5 +9,6 @@
         "noEmit": true,
         "allowJs": true
     },
+    "include": ["src/**/*", "__tests__/**/*"],
     "exclude": ["node_modules", "dist"]
 }
